requirements.txt
python-dotenv==1.0.1
requests==2.31.0
selenium==4.25.0
wget==3.2
